refactor(storage): rename firebase-storage → blob-storage, drop dead getFirebaseApp
The util is 100% @vercel/blob; the firebase-storage name and the
throwing getFirebaseApp() stub were Firebase-era leftovers with zero
live references (no firebase npm dep, stub never invoked).

- git mv server/utils/firebase-storage.ts → blob-storage.ts (3 brands)
- remove dead getFirebaseApp() export + its dead test mocks
  (getFirebaseApp, firebaseStorageBucket) in [slug].delete.test.ts
- update 15 production importers + 4 alquilatucarro test mock paths

Behavior-preserving: storage suite 44/44 (== baseline), typecheck
delta 0 (640/640 errors, 22/22 TS2307 — pre-existing ~/-alias noise).
Firebase Hosting config (firebase.json/.firebaserc) intentionally
left untouched — separate concern.

Holdout: docs/specs/blob-storage-rename/scenarios/ (6/6 satisfied).
